fix: 프리패치 위치 모임방으로 정정 -> 커스텀훅으로 만듦, 헤더 스크롤 감지 로직 커스텀훅으로 분리
