+++ src/app/groups/[id]/_components/GroupSection.tsx
@@ -1,53 +1,30 @@
 'use client';
 
-import { usePathname } from 'next/navigation';
+import { useParams } from 'next/navigation';
 
-import { useEffect, useRef, useState } from 'react';
+import { useEffect, useRef } from 'react';
 
 import Header from '@components/common/Header';
 
 import { GroupCardSection } from '@app/groups/[id]/_components/GroupCardSection';
 import TabPages from '@app/groups/[id]/_components/TabPages';
 
-import useHeaderStore from '@stores/useHeaderStore';
 import useModalStore from '@stores/useModalStore';
+import useManagementPrefetch from '@hooks/group/useManagementPrefetch';
+import useObserveScrollHeader from '@hooks/group/useObserveScrollHeader';
 
 const GroupSection = () => {
-  const [isScrolled, setIsScrolled] = useState<boolean>(false);
+  const { id } = useParams();
+  const groupId = Array.isArray(id) ? id[0] : id;
+  useManagementPrefetch({ groupId });
 
   const targetRef = useRef<HTMLDivElement | null>(null);
-
-  const pathName = usePathname();
-
-  const { reset } = useHeaderStore();
+  const isScrolled = useObserveScrollHeader({targetRef});
 
   const { closeModal } = useModalStore();
 
   useEffect(() => {
     closeModal(); //이전에 열린 모달 닫기 처리용
-
-    const targetInstanceRef = targetRef.current;
-
-    const options = {
-      rootMargin: '-48px 0px 0px 0px',
-      threshold: 0,
-    };
-
-    const observer = new IntersectionObserver((entries) => {
-      setIsScrolled(!entries[0].isIntersecting);
-    }, options);
-
-    if (targetInstanceRef) {
-      observer.observe(targetInstanceRef);
-    }
-
-    return () => {
-      if (targetInstanceRef) {
-        observer.unobserve(targetInstanceRef);
-      }
-
-      if (pathName !== '/') reset();
-    };
   }, []);
 
   return (
